+++ client/src/components/common/Input.vue
@@ -48,7 +48,7 @@ export default {
 };
 </script>
 
-<style>
+<style scoped>
 .label {
   display: block;
   position: relative;
